requirements.txt
# RSS Feed Parsing
feedparser>=6.0.10

# Scheduling
apscheduler>=3.10.4
pytz>=2023.3

# Templating
jinja2>=3.1.2

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.5.0

# HTML Processing
beautifulsoup4>=4.12.2
lxml>=4.9.3

# HTTP Requests
requests>=2.31.0

# Date parsing
python-dateutil>=2.8.2

# String Similarity (for fuzzy deduplication)
python-Levenshtein>=0.23.0

# Gmail API (alternative to SMTP)
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0